refactor(Koshy_branch): Fix ctype of speeds + fix bugs with parallax && animations

+++ src/parallax/set_data/set_layer_speed.c
@@ -9,7 +9,7 @@
 
 void set_layer_speed(parallax_t *back)
 {
-    size_t multiplicator = NB_FILES / 2;
+    double multiplicator = NB_FILES / 2;
 
     if (FILEPATHS == NULL)
         return;
